# openapi.yml
openapi: 3.1.0
info:
  title: MemoryOS browser API
  description: The same-origin contract consumed by the MemoryOS browser application.
  version: 1.0.0
servers:
- url: /
  description: Same-origin MemoryOS runtime
paths:
  /api/sources/{sourceId}/google-drive/schedule:
    put:
      tags:
      - Sources
      summary: Update the automatic Google Drive sync interval
      operationId: updateGoogleDriveSchedule
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoogleDriveScheduleRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveConfigurationResponse"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/sources/{sourceId}/google-drive/roots:
    put:
      tags:
      - Sources
      summary: Replace selected Google Drive roots without changing the creation-time
        scope mode
      operationId: replaceGoogleDriveRoots
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceGoogleDriveRootsRequest"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveSelectionReceiptResponse"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/sources/{sourceId}/google-drive/sync:
    post:
      tags:
      - Sources
      summary: Schedule durable Google Drive synchronization
      operationId: synchronizeGoogleDriveSource
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SourceOperation"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/sources/{sourceId}/google-drive/pause:
    post:
      tags:
      - Sources
      summary: Pause or resume future automatic Google Drive synchronization
      operationId: updateGoogleDrivePause
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoogleDrivePauseRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveConfigurationResponse"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/sources/{sourceId}/google-drive/linked-documents/discover:
    post:
      tags:
      - Sources
      summary: Discover linked documents without approving or synchronizing them
      operationId: discoverGoogleDriveLinkedDocuments
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveConfigurationResponse"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/sources/google-drive:
    post:
      tags:
      - Sources
      summary: Create a Google Drive source using a reusable credential
      operationId: createGoogleDriveSource
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoogleDriveSourceRequest"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveSelectionReceiptResponse"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/sources:
    get:
      tags:
      - Sources
      summary: List Tenant sources
      operationId: listSources
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceSummary"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/sources/{sourceId}/google-drive:
    get:
      tags:
      - Sources
      summary: Get Google Drive source configuration
      operationId: getGoogleDriveConfiguration
      parameters:
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveConfigurationResponse"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/sources/{sourceId}/google-drive/selection:
    get:
      tags:
      - Sources
      summary: Page selected roots and verified linked documents within a pinned Source
        snapshot
      operationId: getGoogleDriveSelection
      parameters:
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: kind
        in: query
        required: false
        schema:
          type: string
          enum:
          - FOLDER
          - FILE
          - LINKED
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 25
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveSelectionResponse"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/sources/{sourceId}/google-drive/selection-tree:
    get:
      tags:
      - Sources
      summary: "Expand selected Google Drive folders, files, and discovered links\
        \ within the Source scope"
      operationId: getGoogleDriveSelectionTree
      parameters:
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: parentId
        in: query
        required: false
        schema:
          type: string
      - name: cursor
        in: query
        required: false
        schema:
          type: string
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 25
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveSelectionTreeResponse"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/sources/{sourceId}/google-drive/selection-draft:
    get:
      tags:
      - Sources
      summary: Get the complete bounded active selection for editing
      operationId: getGoogleDriveSelectionDraft
      parameters:
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveSelectionDraftResponse"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/sources/google-drive/selection-requests/{requestId}:
    get:
      tags:
      - Sources
      summary: Recover an accepted selection receipt for the initiating owner
      operationId: getGoogleDriveSelectionRequest
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveSelectionReceiptResponse"
      security:
      - bearerAuth: []
      - browserSession: []
  /api/sources/google-drive/selection-policy:
    get:
      tags:
      - Sources
      summary: Get the configured selection admission limits
      operationId: getGoogleDriveSelectionPolicy
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveSelectionPolicyResponse"
      security:
      - bearerAuth: []
      - browserSession: []
...
components:
  schemas:
    UpdateGoogleDriveScheduleRequest:
      type: object
      additionalProperties: false
      properties:
        syncIntervalMinutes:
          type: integer
          format: int32
          maximum: 2147483647
          minimum: 1
      required:
      - syncIntervalMinutes
    GoogleDriveConfigurationResponse:
      type: object
      additionalProperties: false
      properties:
        sourceId:
          type: string
          format: uuid
        credentialId:
          type: string
          format: uuid
        accountEmail:
          type: string
        credentialStatus:
          type: string
        credentialRevision:
          type: integer
          format: int64
        oauthClientConfigured:
          type: boolean
        revision:
          type: integer
          format: int64
        syncIntervalMinutes:
          type: integer
          format: int32
          maximum: 2147483647
          minimum: 1
        scheduleRevision:
          type: integer
          format: int64
          minimum: 1
        syncPaused:
          type: boolean
        scopeMode:
          type: string
          enum:
          - GENERAL
          - SPECIFIC
        counts:
          $ref: "#/components/schemas/GoogleDriveSelectionCountsResponse"
        discoveryRevision:
          type: integer
          format: int64
          minimum: 0
        discoveredAt:
          type:
          - string
          - "null"
          format: date-time
        discoveryErrors:
          type: array
          items:
            $ref: "#/components/schemas/GoogleDriveDiscoveryErrorResponse"
        lastSyncedAt:
          type:
          - string
          - "null"
          format: date-time
        pendingWork:
          type: boolean
        errorCode:
          type:
          - string
          - "null"
        pendingSelectionOperation:
          oneOf:
          - $ref: "#/components/schemas/SourceOperation"
          - type: "null"
      required:
      - accountEmail
      - counts
      - credentialId
      - credentialRevision
      - credentialStatus
      - discoveredAt
      - discoveryErrors
      - discoveryRevision
      - errorCode
      - lastSyncedAt
      - oauthClientConfigured
      - pendingSelectionOperation
      - pendingWork
      - revision
      - scheduleRevision
      - scopeMode
      - sourceId
      - syncIntervalMinutes
      - syncPaused
    GoogleDriveDiscoveryErrorResponse:
      type: object
      additionalProperties: false
      properties:
        fileId:
          type: string
        fileName:
          type: string
        code:
          type: string
      required:
      - code
      - fileId
      - fileName
    GoogleDriveSelectionCountsResponse:
      type: object
      additionalProperties: false
      properties:
        folders:
          type: integer
          format: int64
        files:
          type: integer
          format: int64
        linkedDocuments:
          type: integer
          format: int64
        approvedLinkedDocuments:
          type: integer
          format: int64
      required:
      - approvedLinkedDocuments
      - files
      - folders
      - linkedDocuments
    SourceOperation:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type:
          - string
          - "null"
          format: date-time
        errorCode:
          type:
          - string
          - "null"
      required:
      - completedAt
      - createdAt
      - errorCode
      - id
      - status
      - type
    ReplaceGoogleDriveRootsRequest:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          format: uuid
        discoveryRevision:
          type: integer
          format: int64
          minimum: 0
        credentialRevision:
          type: integer
          format: int64
          minimum: 0
        scopeMode:
          type: string
          description: GENERAL includes the connected account's My Drive tree; SPECIFIC
            includes selected links.
          enum:
          - GENERAL
          - SPECIFIC
        links:
          type: array
          description: Empty for GENERAL; distinct non-overlapping links bounded by
            the selection policy for SPECIFIC.
          items:
            type: string
            maxLength: 2048
            minLength: 0
        linkedDocumentIds:
          type: array
          description: Distinct discovered document IDs explicitly approved for ingestion.
            Empty for GENERAL.
          items:
            type: string
            minLength: 1
            pattern: "[A-Za-z0-9_-]{1,256}"
          maxItems: 500
          minItems: 0
          uniqueItems: true
      required:
      - credentialRevision
      - discoveryRevision
      - linkedDocumentIds
      - links
      - requestId
      - scopeMode
    GoogleDriveSelectionReceiptResponse:
      type: object
      additionalProperties: false
      properties:
        sourceId:
          type: string
          format: uuid
        operation:
          $ref: "#/components/schemas/SourceOperation"
      required:
      - operation
      - sourceId
...
    SourceSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        access:
          type: string
        status:
          type: string
        pendingWork:
          type: boolean
        documentCount:
          type: integer
          format: int64
        lastSucceededAt:
          type:
          - string
          - "null"
          format: date-time
        errorCode:
          type:
          - string
          - "null"
        actions:
          type: array
          items:
            type: string
            enum:
            - upload
            - reindex
            - remove_items
            - delete
            - manage_groups
            - rename
            - manage_access
            - manage_configuration
            - synchronize
            - manage_schedule
            - pause_sync
            - resume_sync
      required:
      - access
      - actions
      - documentCount
      - errorCode
      - id
      - lastSucceededAt
      - name
      - pendingWork
      - status
      - type
...
    UpdateGoogleDrivePauseRequest:
      type: object
      additionalProperties: false
      properties:
        expectedRevision:
          type: integer
          format: int64
          minimum: 1
        paused:
          type: boolean
      required:
      - expectedRevision
      - paused
...
    CreateGoogleDriveSourceRequest:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
          minLength: 0
        credentialId:
          type: string
          format: uuid
        scopeMode:
          type: string
          description: GENERAL includes the connected account's My Drive tree; SPECIFIC
            includes selected links.
          enum:
          - GENERAL
          - SPECIFIC
        links:
          type: array
          description: Empty for GENERAL; distinct non-overlapping links bounded by
            the selection policy for SPECIFIC.
          items:
            type: string
            maxLength: 2048
            minLength: 0
        groupIds:
          type:
          - array
          - "null"
          description: Ordinary groups; at least one managed group is required for
            scoped managers. Global creation may omit groups.
          items:
            type: string
            format: uuid
          maxItems: 100
          minItems: 0
      required:
      - credentialId
      - links
      - name
      - requestId
      - scopeMode
...
    GoogleDriveLinkOriginResponse:
      type: object
      additionalProperties: false
      properties:
        rootId:
          type: string
        parentId:
          type: string
        parentName:
          type: string
        location:
          type: string
      required:
      - location
      - parentId
      - parentName
      - rootId
    GoogleDriveSelectionItemResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
        kind:
          type: string
          enum:
          - FOLDER
          - FILE
          - LINKED
        selected:
          type: boolean
        coveredByRoots:
          type: boolean
        status:
          type: string
          enum:
          - AVAILABLE
          - UNAVAILABLE
          - UNSUPPORTED
        origins:
          type: array
          items:
            $ref: "#/components/schemas/GoogleDriveLinkOriginResponse"
      required:
      - coveredByRoots
      - id
      - kind
      - mimeType
      - name
      - origins
      - selected
      - status
    GoogleDriveSelectionResponse:
      type: object
      additionalProperties: false
      properties:
        revision:
          type: integer
          format: int64
        discoveryRevision:
          type: integer
          format: int64
        credentialRevision:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/GoogleDriveSelectionItemResponse"
        nextCursor:
          type:
          - string
          - "null"
        counts:
          $ref: "#/components/schemas/GoogleDriveSelectionCountsResponse"
      required:
      - counts
      - credentialRevision
      - discoveryRevision
      - items
      - nextCursor
      - revision
    GoogleDriveSelectionTreeItemResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
        kind:
          type: string
          enum:
          - FOLDER
          - FILE
          - LINKED
        selected:
          type: boolean
        coveredByRoots:
          type: boolean
        status:
          type: string
          enum:
          - AVAILABLE
          - UNAVAILABLE
          - UNSUPPORTED
        origins:
          type: array
          items:
            $ref: "#/components/schemas/GoogleDriveLinkOriginResponse"
        expandable:
          type: boolean
      required:
      - coveredByRoots
      - expandable
      - id
      - kind
      - mimeType
      - name
      - origins
      - selected
      - status
    GoogleDriveSelectionTreeResponse:
      type: object
      additionalProperties: false
      properties:
        revision:
          type: integer
          format: int64
        discoveryRevision:
          type: integer
          format: int64
        credentialRevision:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: "#/components/schemas/GoogleDriveSelectionTreeItemResponse"
        nextCursor:
          type:
          - string
          - "null"
        counts:
          $ref: "#/components/schemas/GoogleDriveSelectionCountsResponse"
      required:
      - counts
      - credentialRevision
      - discoveryRevision
      - items
      - nextCursor
      - revision
    GoogleDriveSelectionDraftResponse:
      type: object
      additionalProperties: false
      properties:
        revision:
          type: integer
          format: int64
        discoveryRevision:
          type: integer
          format: int64
        credentialRevision:
          type: integer
          format: int64
        links:
          type: array
          items:
            type: string
        linkedDocumentIds:
          type: array
          items:
            type: string
      required:
      - credentialRevision
      - discoveryRevision
      - linkedDocumentIds
      - links
      - revision
...
    GoogleDriveSelectionPolicyResponse:
      type: object
      additionalProperties: false
      properties:
        maxExplicitRootsPerSource:
          type: integer
          format: int32
        maxRequestBytes:
          type: integer
          format: int32
        maxLinkedDocuments:
          type: integer
          format: int32
      required:
      - maxExplicitRootsPerSource
      - maxLinkedDocuments
      - maxRequestBytes
...
  securitySchemes:
    browserSession:
      type: apiKey
      description: HttpOnly JDBC-backed Spring Security session cookie.
      name: SESSION
      in: cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT